Make build script robust for Render

+++ build.sh
@@ -1,6 +1,11 @@
 #!/usr/bin/env bash
 set -o errexit
 
-cd server
-pip install -r requirements.txt
-python manage.py collectstatic --noinput
+if [ -f "requirements.txt" ]; then
+  pip install -r requirements.txt
+  python manage.py collectstatic --noinput
+else
+  cd server
+  pip install -r requirements.txt
+  python manage.py collectstatic --noinput
+fi
